requirements.txt
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
dialogflow==0.5.1
dj-database-url==0.5.0
Django==2.1.2
djangorestframework==3.9.0
future==0.17.0
google-api-core==1.5.1
google-auth==1.5.1
googleapis-common-protos==1.5.5
grpcio==1.16.0
gunicorn==19.9.0
httpie==0.9.9
idna==2.7
line-bot-sdk==1.8.0
protobuf==3.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pytz==2018.7
requests==2.20.0
rsa==4.0
six==1.11.0
urllib3==1.24
whitenoise==4.1
